+++ src/JustSaying/Messaging/MessageSerialization/NewtonsoftSerializationFactory.cs
@@ -3,7 +3,7 @@ using Newtonsoft.Json;
 
 namespace JustSaying.Messaging.MessageSerialization;
 
-public class NewtonsoftSerializationFactory(JsonSerializerSettings settings = null) : IMessageBodySerializationFactory
+public sealed class NewtonsoftSerializationFactory(JsonSerializerSettings settings = null) : IMessageBodySerializationFactory
 {
     public IMessageBodySerializer GetSerializer<T>() where T : Message => new NewtonsoftMessageBodySerializer<T>(settings);
 }
