+++ src/JustSaying/Messaging/MessageSerialization/SystemTextJsonSerializer`1.cs
