+++ tests/JustSaying.IntegrationTests/Fluent/IntegrationTestBase.cs
@@ -1,8 +1,5 @@
 using System.Diagnostics;
-using Amazon;
 using Amazon.Runtime;
-using Amazon.SimpleNotificationService;
-using Amazon.SQS;
 using JustSaying.AwsTools;
 using JustSaying.Messaging;
 using JustSaying.Messaging.MessageHandling;
@@ -137,24 +134,5 @@ public abstract class IntegrationTestBase(ITestOutputHelper outputHelper)
         }
 
         await actionTask;
-    }
-}
-
-public sealed class LocalAwsClientFactory : IAwsClientFactory
-{
-    private readonly InMemoryAwsBus _bus;
-
-    public LocalAwsClientFactory(InMemoryAwsBus bus)
-    {
-        _bus = bus;
-    }
-    public IAmazonSimpleNotificationService GetSnsClient(RegionEndpoint region)
-    {
-        return _bus.CreateSnsClient();
-    }
-
-    public IAmazonSQS GetSqsClient(RegionEndpoint region)
-    {
-        return _bus.CreateSqsClient();
     }
 }
