+++ tests/JustSaying.TestingFramework/LocalAwsClientFactory.cs
@@ -1,36 +1,26 @@
 using Amazon;
-using Amazon.Runtime;
 using Amazon.SimpleNotificationService;
 using Amazon.SQS;
 using JustSaying.AwsTools;
+using LocalSqsSnsMessaging;
 
 namespace JustSaying.TestingFramework;
 
-// public class LocalAwsClientFactory(Uri serviceUrl) : IAwsClientFactory
-// {
-//     private Uri ServiceUrl { get; } = serviceUrl;
-//
-//     public IAmazonSimpleNotificationService GetSnsClient(RegionEndpoint region)
-//     {
-//         var credentials = new AnonymousAWSCredentials();
-//         var clientConfig = new AmazonSimpleNotificationServiceConfig
-//         {
-//             RegionEndpoint = region,
-//             ServiceURL = ServiceUrl.ToString()
-//         };
-//
-//         return new AmazonSimpleNotificationServiceClient(credentials, clientConfig);
-//     }
-//
-//     public IAmazonSQS GetSqsClient(RegionEndpoint region)
-//     {
-//         var credentials = new AnonymousAWSCredentials();
-//         var clientConfig = new AmazonSQSConfig
-//         {
-//             RegionEndpoint = region,
-//             ServiceURL = ServiceUrl.ToString()
-//         };
-//
-//         return new AmazonSQSClient(credentials, clientConfig);
-//     }
-// }
+public sealed class LocalAwsClientFactory : IAwsClientFactory
+{
+    private readonly InMemoryAwsBus _bus;
+
+    public LocalAwsClientFactory(InMemoryAwsBus bus)
+    {
+        _bus = bus;
+    }
+    public IAmazonSimpleNotificationService GetSnsClient(RegionEndpoint region)
+    {
+        return _bus.CreateSnsClient();
+    }
+
+    public IAmazonSQS GetSqsClient(RegionEndpoint region)
+    {
+        return _bus.CreateSqsClient();
+    }
+}
